Update smax_env.py
Start positions are currently hard-coded to assume that the map_width and height is 32. If the map width is set to 128, the units will this start in the top-left corner.

Make team_0 and team_1 start y coordinates halfway up the map_height (instead of hard-coded 16). The x coordinate will be a fourth of the map_width, and three fourth of the map_width respectively.